+++ crates/stackable-versioned-macros/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "stackable-versioned-macros"
+version = "0.1.0"
+authors.workspace = true
+license.workspace = true
+edition.workspace = true
+repository.workspace = true
+
+[lib]
+proc-macro = true
+
+[dependencies]
+k8s-version = { path = "../k8s-version", features = ["darling"] }
+
+darling.workspace = true
+proc-macro2.workspace = true
+syn.workspace = true
+quote.workspace = true
+
+[dev-dependencies]
+rstest.workspace = true
