+++ crates/stackable-versioned-macros/src/gen/neighbors.rs
