+++ crates/stackable-versioned-macros/src/gen/version.rs
@@ -1,7 +1,9 @@
 use k8s_version::Version;
+use syn::Ident;
 
 #[derive(Debug, Clone)]
 pub(crate) struct ContainerVersion {
     pub(crate) deprecated: bool,
     pub(crate) inner: Version,
+    pub(crate) ident: Ident,
 }
