+++ crates/stackable-versioned/Cargo.toml
@@ -6,16 +6,5 @@ license.workspace = true
 edition.workspace = true
 repository.workspace = true
 
-[lib]
-proc-macro = true
-
 [dependencies]
-k8s-version = { path = "../k8s-version", features = ["darling"] }
-
-darling.workspace = true
-proc-macro2.workspace = true
-syn.workspace = true
-quote.workspace = true
-
-[dev-dependencies]
-rstest.workspace = true
+stackable-versioned-macros = { path = "../stackable-versioned-macros" }
